made a new SCWindow class base for all windows for easier usage and incorperated
