+++ keylog.txt
@@ -1,3 +1,2 @@
 ***KEYLOGGING FILE***
-(Delivered at DAY MON DAYNUM HOUR:MIN:SEC YEAR: ['e', 'x', 'a', 'm', 'p', 'l', 'e', 'ctrl'], example),
-(Delivered at Fri Jul  4 12:23:15 2025: ['ctrl'], )
+(Delivered at DAY MON DT HR:MN:SC YEAR: ['e', 'x', 'a', 'm', 'p', 'l', 'e', 'ctrl'], example),
